+++ Commons/src/main/java/org/jmangos/commons/model/container/AccountsContainer.java
@@ -11,14 +11,14 @@ public class AccountsContainer extends IdNameKeyContainer<AccountInfo> {
     private static final Logger logger = LoggerFactory.getLogger(AccountsContainer.class);
     
     @Override
-    public void onDataNotPresentEvent(final int objectId) {
+    public void onDataNotPresentEvent(final long objectId) {
     
         logger.error("Account with id {} not exist in container.", objectId);
         
     }
     
     @Override
-    public void onDataAlreadyAddedEvent(final int objectId) {
+    public void onDataAlreadyAddedEvent(final long objectId) {
     
         logger.error("Account data already in container. Duplicate Id : {}", objectId);
